+++ src/main/java/com/someguyssoftware/ddenizens/datagen/DDItemTagsProvider.java
@@ -6,6 +6,7 @@ import com.someguyssoftware.ddenizens.tags.DDTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import org.jetbrains.annotations.Nullable;
@@ -20,6 +21,19 @@ public class DDItemTagsProvider extends ItemTagsProvider {
 
     @Override
     protected void addTags(HolderLookup.Provider provider) {
-        tag(DDTags.Items.EGGS).add(Registration.BEHOLDER_EGG.get());
+//        tag(DDTags.Items.EGGS).add(Registration.BEHOLDER_EGG.get());
+
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_AXE1.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_AXE2.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_SWORD1.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_SWORD2.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_SWORD3.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Registration.RUSTY_IRON_SWORD4.get());
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.GOLDEN_SWORD);
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.GOLDEN_AXE);
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.IRON_SWORD);
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.IRON_AXE);
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.STONE_SWORD);
+        tag(DDTags.Items.SKELETON_WARRIOR_WEAPONS).add(Items.STONE_AXE);
     }
 }