+++ src/main/java/com/someguyssoftware/ddenizens/item/DDEggItem.java
@@ -1,3 +1,22 @@
+/*
+ * This file is part of  Dungeon Denizens.
+ * Copyright (c) 2024 Mark Gottschling (gottsch)
+ *
+ * All rights reserved.
+ *
+ * Dungeon Denizens is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Dungeon Denizens is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Dungeon Denizens.  If not, see <http://www.gnu.org/licenses/lgpl>.
+ */
 package com.someguyssoftware.ddenizens.item;
 
 import net.minecraft.network.chat.Component;
@@ -12,6 +31,10 @@ import net.minecraftforge.common.ForgeSpawnEggItem;
 import java.util.List;
 import java.util.function.Supplier;
 
+/**
+ * @author Mark Gottschling on Jan 17, 2024
+ *
+ */
 public class DDEggItem extends ForgeSpawnEggItem {
 
     public DDEggItem(Supplier<? extends EntityType<? extends Mob>> type, int backgroundColor, int highlightColor, Properties props) {