+++ src/main/java/com/someguyssoftware/ddenizens/setup/CommonSetup.java
@@ -77,6 +77,7 @@ public class CommonSetup {
 		event.put(Registration.SHADOW_ENTITY_TYPE.get(), Shadow.createAttributes().build());
 		event.put(Registration.SHADOWLORD_ENTITY_TYPE.get(), Shadowlord.createAttributes().build());
 		event.put(Registration.DAEMON_ENTITY_TYPE.get(), Daemon.createAttributes().build());
+		event.put(Registration.SKELETON_WARRIOR_TYPE.get(), SkeletonWarrior.createAttributes().build());
 
 	}
 
@@ -94,6 +95,8 @@ public class CommonSetup {
 		event.register(Registration.GAZER_ENTITY_TYPE.get(), SpawnPlacements.Type.NO_RESTRICTIONS, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, DenizensMonster::checkIsNetherSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
 		event.register(Registration.SPECTATOR_TYPE.get(), SpawnPlacements.Type.NO_RESTRICTIONS, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, DenizensMonster::checkIsNetherSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
 		event.register(Registration.DAEMON_ENTITY_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, DenizensMonster::checkIsNetherSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
+		event.register(Registration.SKELETON_WARRIOR_TYPE.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, DenizensMonster::checkDDMonsterSpawnRules, SpawnPlacementRegisterEvent.Operation.OR);
+
 	}
 
 	@SubscribeEvent
@@ -111,71 +114,24 @@ public class CommonSetup {
 			event.accept(Registration.GAZER_EGG.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
 			event.accept(Registration.SPECTATOR_EGG.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
 			event.accept(Registration.DAEMON_EGG.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.SKELETON_WARRIOR_EGG.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+
 		}
 		else if (event.getTabKey() == CreativeModeTabs.COMBAT) {
 			event.accept(Registration.CLUB.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
 			event.accept(Registration.SPIKED_CLUB.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_AXE1.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_AXE2.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_SWORD1.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_SWORD2.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_SWORD3.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+			event.accept(Registration.RUSTY_IRON_SWORD4.get(), TabVisibility.PARENT_AND_SEARCH_TABS);
+
 		}
 	}
 	
 	@Mod.EventBusSubscriber(modid = DD.MODID, bus = EventBusSubscriber.Bus.FORGE)
 	public static class ForgeBusSucscriber {
-		/*
-		 * Register the Features with Biomes
-		 */
-//		@SubscribeEvent
-//		public static <BiomeLoadingEvent> void onBiomeLoading(final BiomeLoadingEvent event) {
-//			DD.LOGGER.debug("event for biome -> {}, category -> {}", event.getName(), event.getCategory().getName());
-//			/*
-//			 * register mob spawns to biomes
-//			 */
-//			ResourceLocation biome = event.getName();
-//
-//			Registration.ALL_MOBS.forEach(entityType -> {
-//				IMobConfig config = Config.Mobs.MOBS.get(((EntityType<?>)entityType.get()).getRegistryName());
-//
-//				if (config.getSpawnConfig().enable.get()) {
-//					Result result = isBiomeAllowed(biome, event.getCategory(), config.getSpawnConfig());
-//					if (result == Result.OK || result == Result.WHITE_LISTED) {
-//						DD.LOGGER.debug("registering spawner data -> {}", ((EntityType<?>)entityType.get()).getRegistryName());
-//
-//						if (event.getCategory() == BiomeCategory.NETHER && config instanceof INetherMobConfig) {
-//							event.getSpawns().getSpawner(MobCategory.MONSTER)
-//							.add(new MobSpawnSettings.SpawnerData(
-//									(EntityType<?>)entityType.get(),
-//									((INetherMobConfig)config).getNetherSpawn().spawnWeight.get(),
-//									((INetherMobConfig)config).getNetherSpawn().minSpawn.get(),
-//									((INetherMobConfig)config).getNetherSpawn().maxSpawn.get()));
-//						}
-//						else {
-//							event.getSpawns().getSpawner(MobCategory.MONSTER)
-//							.add(new MobSpawnSettings.SpawnerData(
-//									(EntityType<?>)entityType.get(),
-//									config.getSpawnConfig().spawnWeight.get(),
-//									config.getSpawnConfig().minSpawn.get(),
-//									config.getSpawnConfig().maxSpawn.get()));
-//						}
-//					}
-//				}
-//			});
-//		}
-
-		//	    @SubscribeEvent
-		//	    public static void registerParticle(ParticleFactoryRegisterEvent event){
-		//	        Minecraft.getInstance().particleEngine.register(Registration.SHADOW_PARTICLE.get(), ModGlowParticle.BlockOffSideProvider::new);
-		//	    }
-
-		/**
-		 * Helper method
-		 * @param biome
-		 * @param category
-		 * @param config
-		 * @return
-		 */
-//		public static Result isBiomeAllowed(ResourceLocation biome, BiomeCategory category, SpawnConfig config) {
-//			return BiomeHelper.isBiomeAllowed(biome, category, config.biomeWhitelist.get(), config.biomeBlacklist.get(),
-//					config.biomeCategoryWhitelist.get(), config.biomeCategoryBlacklist.get());
-//		}
 
 		@SubscribeEvent
 		public static void addGoals(final EntityJoinLevelEvent event) {