+++ src/main/java/com/someguyssoftware/ddenizens/setup/Registration.java
@@ -22,20 +22,17 @@ import java.util.List;
 import com.google.common.collect.Lists;
 import com.someguyssoftware.ddenizens.DD;
 import com.someguyssoftware.ddenizens.capability.GhoulCapability;
-import com.someguyssoftware.ddenizens.config.Config;
 import com.someguyssoftware.ddenizens.entity.monster.*;
 import com.someguyssoftware.ddenizens.entity.projectile.*;
 
-import com.someguyssoftware.ddenizens.item.BeholderEggItem;
-import com.someguyssoftware.ddenizens.item.DeathTyrantEggItem;
-import com.someguyssoftware.ddenizens.item.GazerEggItem;
-import com.someguyssoftware.ddenizens.item.SpectatorEggItem;
+import com.someguyssoftware.ddenizens.item.*;
 import net.minecraft.core.particles.ParticleType;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.entity.monster.Skeleton;
+import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.SwordItem;
 import net.minecraft.world.item.Tiers;
@@ -167,15 +164,15 @@ public class Registration {
 			.fireImmune()
 			.build(DAEMON));
 
-	public static final RegistryObject<EntityType<Skeleton>> SKELETON_WARRIOR_TYPE = Registration.ENTITIES.register(SKELETON_WARRIOR, () -> EntityType.Builder.of(Skeleton::new, MobCategory.MONSTER)
+	public static final RegistryObject<EntityType<SkeletonWarrior>> SKELETON_WARRIOR_TYPE = Registration.ENTITIES.register(SKELETON_WARRIOR, () -> EntityType.Builder.of(SkeletonWarrior::new, MobCategory.MONSTER)
 			.sized(0.6F, 1.95F)
 			.clientTrackingRange(12)
 			.setShouldReceiveVelocityUpdates(false)
 			.build(SKELETON_WARRIOR));
 	
 	// projectile entities
-	public static final RegistryObject<EntityType<SlowSpell>> SLOW_SPELL_ENTITY_TYPE =
-			Registration.ENTITIES.register(SLOW_SPELL, () -> EntityType.Builder.of(SlowSpell::new, MobCategory.MISC)
+	public static final RegistryObject<EntityType<ParalysisSpell>> SLOW_SPELL_ENTITY_TYPE =
+			Registration.ENTITIES.register(SLOW_SPELL, () -> EntityType.Builder.of(ParalysisSpell::new, MobCategory.MISC)
 			.sized(1F, 1F)
 			.clientTrackingRange(12)
 			.setShouldReceiveVelocityUpdates(false)
@@ -190,14 +187,14 @@ public class Registration {
 
 	public static final RegistryObject<EntityType<DisarmSpell>> DISARM_SPELL_ENTITY_TYPE =
 			Registration.ENTITIES.register(DISARM_SPELL, () -> EntityType.Builder.of(DisarmSpell::new, MobCategory.MISC)
-					.sized(0.5F, 0.75F)
+					.sized(1F, 1F)
 					.clientTrackingRange(12)
 					.setShouldReceiveVelocityUpdates(false)
 					.build(DISARM_SPELL));
 
 	public static final RegistryObject<EntityType<DisintegrateSpell>> DISINTEGRATE_SPELL_ENTITY_TYPE =
 			Registration.ENTITIES.register(DISINTEGRATE_SPELL, () -> EntityType.Builder.of(DisintegrateSpell::new, MobCategory.MISC)
-					.sized(0.5F, 0.75F)
+					.sized(1F, 1F)
 					.clientTrackingRange(12)
 					.setShouldReceiveVelocityUpdates(false)
 					.build(DISINTEGRATE_SPELL));
@@ -221,21 +218,24 @@ public class Registration {
 	 */
 	
 	// mod eggs
-	public static final RegistryObject<Item> HEADLESS_EGG = Registration.ITEMS.register(HEADLESS, () -> new ForgeSpawnEggItem(HEADLESS_ENTITY_TYPE, 0xc8b486, 0x6f5e48, new Item.Properties()));
-	public static final RegistryObject<Item> ORC_EGG = Registration.ITEMS.register(ORC, () -> new ForgeSpawnEggItem(ORC_ENTITY_TYPE, 0xc8b486, 0x6f5e48, new Item.Properties()));
-	public static final RegistryObject<Item> GHOUL_EGG = Registration.ITEMS.register(GHOUL, () -> new ForgeSpawnEggItem(GHOUL_ENTITY_TYPE, 0x93aba3, 0x869e96, new Item.Properties()));
+	public static final RegistryObject<Item> HEADLESS_EGG = Registration.ITEMS.register(HEADLESS + "_egg", () -> new ForgeSpawnEggItem(HEADLESS_ENTITY_TYPE, 0xc8b486, 0x6f5e48, new Item.Properties()));
+	public static final RegistryObject<Item> ORC_EGG = Registration.ITEMS.register(ORC + "_egg", () -> new ForgeSpawnEggItem(ORC_ENTITY_TYPE, 0xc8b486, 0x6f5e48, new Item.Properties()));
+	public static final RegistryObject<Item> GHOUL_EGG = Registration.ITEMS.register(GHOUL + "_egg", () -> new ForgeSpawnEggItem(GHOUL_ENTITY_TYPE, 0x93aba3, 0x869e96, new Item.Properties()));
 
-	public static final RegistryObject<Item> BEHOLDER_EGG = Registration.ITEMS.register(BEHOLDER, () -> new BeholderEggItem(BEHOLDER_ENTITY_TYPE, 0x871e00, 0xc15227, new Item.Properties()));
-	public static final RegistryObject<Item> DEATH_TYRANT_EGG = Registration.ITEMS.register(DEATH_TYRANT, () -> new DeathTyrantEggItem(DEATH_TYRANT_TYPE, 0x86765a, 0xcdc3bb, new Item.Properties()));
+	public static final RegistryObject<Item> BEHOLDER_EGG = Registration.ITEMS.register(BEHOLDER + "_egg", () -> new BeholderEggItem(BEHOLDER_ENTITY_TYPE, 0x871e00, 0xc15227, new Item.Properties()));
+	public static final RegistryObject<Item> DEATH_TYRANT_EGG = Registration.ITEMS.register(DEATH_TYRANT + "_egg", () -> new DeathTyrantEggItem(DEATH_TYRANT_TYPE, 0x86765a, 0xcdc3bb, new Item.Properties()));
 
-	public static final RegistryObject<Item> GAZER_EGG = Registration.ITEMS.register(GAZER, () -> new GazerEggItem(GAZER_ENTITY_TYPE, 0x7a2e2f, 0x63181d, new Item.Properties()));
-	public static final RegistryObject<Item> SPECTATOR_EGG = Registration.ITEMS.register(SPECTATOR, () -> new SpectatorEggItem(SPECTATOR_TYPE, 0x344133, 0xabb685, new Item.Properties()));
+	public static final RegistryObject<Item> GAZER_EGG = Registration.ITEMS.register(GAZER + "_egg", () -> new GazerEggItem(GAZER_ENTITY_TYPE, 0x7a2e2f, 0x63181d, new Item.Properties()));
+	public static final RegistryObject<Item> SPECTATOR_EGG = Registration.ITEMS.register(SPECTATOR + "_egg", () -> new SpectatorEggItem(SPECTATOR_TYPE, 0x344133, 0xabb685, new Item.Properties()));
 
-	public static final RegistryObject<Item> BOULDER_EGG = Registration.ITEMS.register(BOULDER, () -> new ForgeSpawnEggItem(BOULDER_ENTITY_TYPE, 0x747474, 0x8f8f8f, new Item.Properties()));
-	public static final RegistryObject<Item> SHADOW_EGG = Registration.ITEMS.register(SHADOW, () -> new ForgeSpawnEggItem(SHADOW_ENTITY_TYPE, 0x000000, 0x2b2b2b, new Item.Properties()));
-	public static final RegistryObject<Item> SHADOWLORD_EGG = Registration.ITEMS.register(SHADOWLORD, () -> new ForgeSpawnEggItem(SHADOWLORD_ENTITY_TYPE, 0x000000, 0x050831, new Item.Properties()));
-	public static final RegistryObject<Item> DAEMON_EGG = Registration.ITEMS.register(DAEMON, () -> new ForgeSpawnEggItem(DAEMON_ENTITY_TYPE, 0xff0000, 0xfc0000, new Item.Properties()));
-	
+	public static final RegistryObject<Item> BOULDER_EGG = Registration.ITEMS.register(BOULDER + "_egg", () -> new ForgeSpawnEggItem(BOULDER_ENTITY_TYPE, 0x747474, 0x8f8f8f, new Item.Properties()));
+	public static final RegistryObject<Item> SHADOW_EGG = Registration.ITEMS.register(SHADOW + "_egg", () -> new ForgeSpawnEggItem(SHADOW_ENTITY_TYPE, 0x000000, 0x2b2b2b, new Item.Properties()));
+	public static final RegistryObject<Item> SHADOWLORD_EGG = Registration.ITEMS.register(SHADOWLORD + "_egg", () -> new ForgeSpawnEggItem(SHADOWLORD_ENTITY_TYPE, 0x000000, 0x050831, new Item.Properties()));
+	public static final RegistryObject<Item> DAEMON_EGG = Registration.ITEMS.register(DAEMON + "_egg", () -> new DaemonEggItem(DAEMON_ENTITY_TYPE, 0xff0000, 0xfc0000, new Item.Properties()));
+
+	public static final RegistryObject<Item> SKELETON_WARRIOR_EGG = Registration.ITEMS.register(SKELETON_WARRIOR + "_egg", () -> new ForgeSpawnEggItem(SKELETON_WARRIOR_TYPE, 0xf5f6d2, 0xcdc3bb, new Item.Properties()));
+
+
 	// projectiles
 	public static final RegistryObject<Item> SLOW_SPELL_ITEM = Registration.ITEMS.register(SLOW_SPELL, () -> new Item(new Item.Properties()));
 	public static final RegistryObject<Item> HARM_SPELL_ITEM = Registration.ITEMS.register(HARM_SPELL, () -> new Item(new Item.Properties()));
@@ -252,6 +252,13 @@ public class Registration {
 	public static final RegistryObject<Item> CLUB = Registration.ITEMS.register("club", () -> new SwordItem(Tiers.WOOD, 4, -3.0F, new Item.Properties()));
 	// spiked club is equal to an iron sword but slower
 	public static final RegistryObject<Item> SPIKED_CLUB = Registration.ITEMS.register("spiked_club", () -> new SwordItem(Tiers.WOOD, 5, -3.0F, new Item.Properties()));
+
+	public static final RegistryObject<Item> RUSTY_IRON_SWORD1 = Registration.ITEMS.register("rusty_iron_sword", () -> new SwordItem(Tiers.IRON, 3, -2.4F, new Item.Properties()));
+	public static final RegistryObject<Item> RUSTY_IRON_SWORD2 = Registration.ITEMS.register("rusty_iron_sword_2", () -> new SwordItem(Tiers.IRON, 3, -2.4F, new Item.Properties()));
+	public static final RegistryObject<Item> RUSTY_IRON_SWORD3 = Registration.ITEMS.register("rusty_iron_sword_3", () -> new SwordItem(Tiers.IRON, 3, -2.4F, new Item.Properties()));
+	public static final RegistryObject<Item> RUSTY_IRON_SWORD4 = Registration.ITEMS.register("rusty_iron_sword_4", () -> new SwordItem(Tiers.IRON, 3, -2.4F, new Item.Properties()));
+	public static final RegistryObject<Item> RUSTY_IRON_AXE1 = Registration.ITEMS.register("rusty_iron_axe", () -> new AxeItem(Tiers.IRON, 6F, -3F, new Item.Properties()));
+	public static final RegistryObject<Item> RUSTY_IRON_AXE2 = Registration.ITEMS.register("rusty_iron_axe_2", () -> new AxeItem(Tiers.IRON, 6F, -3F, new Item.Properties()));
 
 	public static final RegistryObject<SoundEvent> AMBIENT_DAEMON = registerSoundEvent("ambient_daemon");
 	public static final RegistryObject<SoundEvent> AMBIENT_BEHOLDER = registerSoundEvent("ambient_beholder");
@@ -272,7 +279,7 @@ public class Registration {
 		ALL_MOBS.add(SHADOWLORD_ENTITY_TYPE);
 		ALL_MOBS.add(DAEMON_ENTITY_TYPE);
 		ALL_MOBS.add(ORC_ENTITY_TYPE);
-//		ALL_MOBS.add(SKELETON_WARRIOR_TYPE);
+		ALL_MOBS.add(SKELETON_WARRIOR_TYPE);
 	}
 	
 	/**