+++ src/main/java/com/someguyssoftware/ddenizens/tags/DDTags.java
@@ -10,6 +10,7 @@ public class DDTags {
 
     public static class Items {
         public static final TagKey<Item> EGGS = mod(DD.MODID, "eggs");
+        public static final TagKey<Item> SKELETON_WARRIOR_WEAPONS = mod(DD.MODID, "skeleton_warrior_weapons");
 
         public static TagKey<Item> mod(String domain, String path) {
             return ItemTags.create(new ResourceLocation(domain, path));